WiimoteReal: Linux: Listen for incoming connections
This adds a scanner which listens for incoming connections from
Wiimotes. It doesn't replace the current scanner, because this only
works for paired Wiimotes, and this isn't very user friendly
(this needs bluetoothd to be disabled and two commands to be run once)

But it does make connecting Wiimotes a bit more similar to the Wii
(Wiimotes can reconnect on any button press, and they can be
disconnected by games).

If we don't have cap_net_bind_service, this scanner will not be able to
do anything.

Unfortunately, this is Linux only because it is not possible to use
L2CAP on Windows, and I don't have the necessary hardware to try
this on OS X.

UGLY/TODO: Change the way the Wiimote scanner works so we don't need to
wait for Bluetooth discovery to finish and so we don't have to poll…