IOS/ES: Expose all other ioctlvs
This exposes all ES ioctlvs to avoid duplicating IOS code everywhere,
to make it easier to reuse (since it's not tied to the PPC IPC
mechanism anymore) and unit test (in the future).

And since we are adding new functions, we may as well give them
proper, official names now, for consistency and not having to guess
which function corresponds to an official name.

I had considered another way of doing this -- adding support for
internal IPC requests inside of IOS -- but decided against it because
having to construct ioctlv requests isn't fun and much more verbose.