[DO NOT MERGE] IOS: Generate the IOS version from the title version
DO NOT MERGE -- Causes a crash for WADs because of the NANDContentManager mess

This is how ES actually does it. Fixes issue 10319 for people who have
system titles (at least IOS TMDs) in their NAND.

Unfortunately, because we still have to support users without a proper
NAND setup, we have to handle missing IOS TMDs gracefully and fall back
to the hardcoded version to make things work in most cases.