Remove GetWiiFSPath
This removes GetWiiFSPath and the associated "open Wii folder"
because it assumes the title data path belongs on the host filesystem
(won't be true with NAND images) and we should not encourage users
to mess with the title data directory manually. We already offer
save import/export features (including bulk export for backups)
so there is little reason to manipulate the emulated NAND directly.

We should make it harder for users to put garbage in the NAND or
manipulate files they should not be touching. Tons of issues that
are reported on the forums are caused by users thinking they can
freely modify / overwrite / copy / move files in the NAND and making
a complete mess afterwards.