IOS/USB: Prevent concurrent device scans
Even though libusb is supposed to be thread-safe, in practice it is not
(at least on Windows); getting a list of devices from two different
threads can result in libusb crashes. This adds a mutex to ensure that
device scanning only happens on a single thread for a given context.