WiiRoot: Add a callback system for FS init/cleanup
Ideally, WiiRoot shouldn't have to know about Movie and NetPlay.
This adds a callback system for Wii FS inits/cleanups that Movie and
NetPlay (and any other subsystem) can use to do whatever they need to
in the Wii root without that code having to be part of WiiRoot.
This new system will be used in future commits.

This probably also removes the need to expose some Movie and NetPlay
internal state but I'll leave that cleanup for another time.